refactor(app): clean project
